Host name filter support
Filter support makes use of Javascript's regex functionality

CSV export includes the hosts that match the filter, matching the hosts
that are displayed.

Popup dialog refreshes are minimized by caching and comparing the host entries
to be displayed. DOM changes are only applied if the host entries differ from the
cached value.